Allow for duplicate mtaQueueId
These postfix generated ids are only guaranteed to be unique
within the same host and at one point in time.

This fix tries to match with recipient email and then by date, latest
first. Currently the logic is only used to set delivery status, hence
this logic should be sufficient.