Add 5.7.26 (DMARC failed)
This error is about as bad as it can get for Tachikoma. This implies
that some or all of the emails sent are being rejected because of invalid
DMARC policy(e.g. incorrect DKIM or SPF policy).

This should be fixed ASAP as this will start earning the sending server a bad rep
and all mails are being hard-bounced.